# model/openapi.json
{"openapi":"3.1.0","info":{"title":"Prophet Revenue Forecast API","description":"\nA local API server for daily revenue forecasting using **Facebook Prophet**.\n\n## Quick workflow\n\n1. `GET  /model/info`       → check what data the model was last trained on\n2. `POST /retrain`          → upload new CSV data to retrain the model\n3. `POST /predict`          → generate a forecast for the next N months\n4. `GET  /jobs/{job_id}`    → poll for job progress\n\n## Async jobs\n\n`/retrain` and `/predict` are **non-blocking** — they return a `job_id` immediately\nand run in the background. Poll `GET /jobs/{job_id}` until `status` is `\"done\"` or `\"error\"`.\n\n## Output files\n\nAll prediction outputs (charts + CSVs) are saved to `outputs/predict_{short_id}/` on disk.\nThe exact path is returned in `result.output_folder` when the job is done.\n","version":"1.0.0"},"paths":{"/model/info":{"get":{"tags":["Model Info"],"summary":"Get current model info","description":"Returns metadata about the currently loaded Prophet model: the **last date it was trained on**, the model file path, the number of training rows, and the timestamp it was trained.","operationId":"model_info_model_info_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelInfo"}}}}}}},"/retrain":{"post":{"tags":["Retrain"],"summary":"Retrain model with new data","description":"Upload a **CSV file** (no header) with two columns:\n- Column 1 – date in `YYYY-MM-DD` format\n- Column 2 – daily revenue as a **raw integer** (same unit as `data.csv`)\n\nThe new data is **merged with the existing history** (duplicates removed by date),\na fresh Prophet model is fitted, and saved under:\n\n```\ntrained_models/{last_date_of_data}/prophet_model.json\n```\n\n`/model/info` and future `/predict` calls will automatically use the new model.\n\nReturns a `job_id` immediately. Poll `GET /jobs/{job_id}` for progress.","operationId":"retrain_retrain_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_retrain_retrain_post"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobStatus"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/predict":{"post":{"tags":["Predict"],"summary":"Start a revenue forecast","description":"Loads the **current Prophet model** and generates a forecast for the next **N months**\n(default `3`). Runs in the background.\n\nPoll `GET /jobs/{job_id}` until `status` is `\"done\"` or `\"error\"`.\n\nWhen done, `result.output_folder` contains all generated files:\n\n| File | Description |\n|---|---|\n| `forecast_daily.csv` | Day-level predictions with confidence interval |\n| `forecast_monthly.csv` | Monthly aggregated totals |\n| `forecast_quarterly.csv` | Quarterly aggregated totals |\n| `forecast_yearly.csv` | Yearly aggregated totals |\n| `forecast_plot.png` | Full timeline chart |\n| `components_plot.png` | Trend / holiday / seasonality breakdown |\n| `monthly_bar.png` | Bar chart of monthly totals |\n\n`result.summary` also returns a JSON summary of the totals so you don't need to open files just for key numbers.","operationId":"predict_predict_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictRequest"}}}},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobStatus"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs/{job_id}":{"get":{"tags":["Jobs"],"summary":"Get job status","description":"Poll this endpoint to track the progress of an async **retrain** or **predict** job. Possible `status` values: `pending` → `running` → `done` | `error`. When `status` is `done`, the `result` field contains the output details.","operationId":"get_job_jobs__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobStatus"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs":{"get":{"tags":["Jobs"],"summary":"List all jobs","description":"Returns all jobs submitted in the **current server session**. The job list is in-memory and resets on server restart.","operationId":"list_jobs_jobs_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/JobStatus"},"type":"array","title":"Response List Jobs Jobs Get"}}}}}}}},"components":{"schemas":{"Body_retrain_retrain_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File","description":"CSV (no header): date, daily_revenue"}},"type":"object","required":["file"],"title":"Body_retrain_retrain_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"JobStatus":{"properties":{"job_id":{"type":"string","title":"Job Id"},"type":{"type":"string","title":"Type"},"status":{"type":"string","enum":["pending","running","done","error"],"title":"Status"},"message":{"type":"string","title":"Message"},"result":{"anyOf":[{},{"type":"null"}],"title":"Result"}},"type":"object","required":["job_id","type","status","message"],"title":"JobStatus"},"ModelInfo":{"properties":{"last_trained_date":{"type":"string","title":"Last Trained Date"},"model_filename":{"type":"string","title":"Model Filename"},"training_rows":{"type":"integer","title":"Training Rows"},"trained_at":{"type":"string","title":"Trained At"}},"type":"object","required":["last_trained_date","model_filename","training_rows","trained_at"],"title":"ModelInfo"},"PredictRequest":{"properties":{"months":{"type":"integer","title":"Months","default":3}},"type":"object","title":"PredictRequest","example":{"months":3}},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}